Refactor tutorials page
