+++ web/src/pages/TutorialsPage/components/ErrorComponent.tsx
@@ -1,0 +1,7 @@
+export default function ErrorComponent({error}:{error:string}){
+  return (
+    <div className="text-center py-20">
+      <p className="text-destructive">{error}</p>
+    </div>
+  );
+}
