+++ web/src/pages/TutorialsPage/components/Loading.tsx
@@ -1,0 +1,7 @@
+export default function Loading(){
+  return (
+    <div className="flex items-center justify-center py-20">
+      <div className="animate-spin rounded-full h-6 w-6 border-2 border-primary border-t-transparent"></div>
+    </div>
+  )
+}
